requirements.txt
django
django-allauth
mysqlclient
whitenoise==5.0.1
django-userforeignkey
pyjwt
speechrecognition
pydub